requirements.txt
Django==3.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.1  # Драйвер для PostgreSQL
django-cors-headers==3.9.0  # Для обработки CORS 
django-filter==2.4.0  # Для фильтрации запросов в Django REST framework
django-environ==0.6.0  # Для управления переменными окружения
djangorestframework-simplejwt==5.2.2
Pillow==8.3.2